Sessions, refactors.
I created Sessions for the app to allow data to be held on the users browser and speed up the app.

            var thisSession = HttpContext.Session.GetObjectFromJson<SessionModel>("thisSession");

The posted command line will allow you to create a SessionModel anywhere and thus access the data within.  For example, place it in a controller and you can pass it to that view.

You can also use it directly within a view if controller method is there.

I refactored some GUI paddings and containers to uniform a little more.

I removed all placeholders I could find.

I added home button to allow users back upon log out.